requirement.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-filter==21.1
django-heroku==0.3.1
django-matplotlib==0.1
gunicorn==20.1.0
joblib==1.1.0
numpy==1.22.2
pandas==1.4.0
Pillow==9.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2021.5
whitenoise==5.3.0
xlrd==2.0.1
